TDQ-13538 TDQ-13267 API: Divide StringConverter into 2 class(StringTrimmer and DuplicateCharEraser).
